+++ src/plooney81/nectar/jsql.clj
@@ -3,6 +3,7 @@
            (net.sf.jsqlparser.parser CCJSqlParserUtil)
            (net.sf.jsqlparser.statement.create.table ColDataType)
            (net.sf.jsqlparser.statement.insert Insert)
+           (net.sf.jsqlparser.statement.update Update)
            (net.sf.jsqlparser.statement.select
              FromItem GroupByElement Join Limit Offset OrderByElement PlainSelect SelectItem SetOperationList WithItem)
            (net.sf.jsqlparser.expression
@@ -190,6 +191,9 @@
 
 (defmethod get-order-by-elements AnalyticExpression [jsql-expr]
   (.getOrderByElements jsql-expr))
+
+(defmethod get-order-by-elements Update [jsql-update]
+  (.getOrderByElements jsql-update))
 
 (defmethod get-order-by-elements :default [jsql-expr]
   (throw (IllegalArgumentException.