+++ src/plooney81/nectar/sql/impl.clj
@@ -20,6 +20,12 @@
     (.getClass jsql-insert)))
 
 (defmulti
+  update->honey
+  "Converts jsql-update objects to honey-sql based on the class"
+  (fn [_honey jsql-update]
+    (.getClass jsql-update)))
+
+(defmulti
   expression
   "Converts a jsql expression to an alternative syntax that will ultimately be used to convert to honeysql"
   (fn [jsql-expr]